Obj position (#14)
* dev commit

* feat: зона вокруг дрона, в которой просчитывается траектория объектра;
feat: минимальная и максимальная дистанция в launch файле;
feat: сброс внетренного состяния с помощью aruco.

* feat: добавлена фильтрация дистанциия до объекта;
feat: адаптированы фильтры Гайвера для ROS;
feat: можно собрать balltracker с фильтром Калмана или "бегущее среднее";
fix: исправлен путь для маркера дрона в rviz;
fix: при анализе дистанций проверяет меньше точек за счет проверки только области интереса. Меньше копирований;
feat: в rviz подсвечивается зона в которой будет вычисляться траектория объекта.

+++ src/puzzle_reaction/include/puzzle_reaction/vision/BallTracking.hpp
@@ -33,7 +33,7 @@ public:
    * @param[out] center center of target color area
    * @param[out] radius area radius
    */
-  void process(cv::Mat &color, cv::Mat &mask, cv::Point2i *center = nullptr, uint16_t *radius = nullptr);
+  void process(cv::Mat &color, cv::Mat &mask, cv::Point2f *center = nullptr, uint16_t *radius = nullptr);
 
   inline void setThreshold(const threshold_t th) { _threshold = th; }
 };
